funçao de criar csv modularizada e testada
