+++ twitter/TwitterToolkit.py
@@ -35,19 +35,6 @@ class TwitterAPI(tweepy.API):
 
         return tweet_list
 
-    def create_csv_basic():
-        with open('twitter/autores.csv', 'w') as csvfile:
-            writer_t = csv.writer(csvfile, delimiter=';')
-            writer_t.writerow(["nome", "seguidores", "tweets", "seguindo", "curtidas"])
-            csvfile.close()
-
-    def update_csv_new_autors(self):
-        with open('twitter/autores.csv', 'a') as csvfile:
-            writer_t = csv.writer(csvfile, delimiter=';')
-            writer_t.writerow([self.name, self.followers_count,
-            self.tweets_count, self.following_count, self.likes_count])
-            csvfile.close()
-
 
 class TweetTK:
