+++ twitter/autores.csv
@@ -1,0 +1,1 @@
+nome;seguidores;tweets;seguindo;curtidas
